bugfix: 클라이언트 컴포넌트에서 supabse server client 사용해 발생하는 문제 해결
